Fixed column types

+++ src/main/resources/db/migration/v0.26.0.2__add_ontology_counts.sql
@@ -1,7 +1,7 @@
 ALTER TABLE ontologyterm
-   ADD COLUMN childcount bigint,
+   ADD COLUMN childcount int,
    ADD COLUMN descendentcount int;
 
 ALTER TABLE ontologyterm_aud
-   ADD COLUMN childcount bigint,
+   ADD COLUMN childcount int,
    ADD COLUMN descendentcount int;
